backend/conversational-ai-gemini-twilio/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fluffyduck-gemini-twilio"
version = "0.1.0"
description = "A bridge between Twilio Voice Calls and Google Gemini AI for real-time voice conversations"
readme = "README.md"
requires-python = ">=3.7"
license = {text = "MIT"}
authors = [
    {name = "FluffyDuck Team"}
]
dependencies = [
    "google-cloud-genai>=0.1.0",
    "quart>=0.18.4",
    "python-dotenv>=1.0.0",
    "twilio>=8.10.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
    "ruff>=0.0.0"
]

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ["py37"]

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.7"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.ruff]
line-length = 88
target-version = "py37"
select = ["E", "F", "B", "I"]
ignore = ["E501"] 